move _to_np_mode and _to_ndimage_mode to _shared.utils.py
